Reworked the Helper slot, tidied code
The helper slot, used to render the currently dragged icon on top of all UI elements, has been moved. It sits at the bottom of the Canvas hierarchy and is referenced by all slots on Start().

Tidied up code in DragAndDrop.cs to reduce the amount of Find() calls when a swap has been initiated. Reduced Find() calls from 4 to 2, and GetComponent<> calls from 6 to 4.