Interactable objects, UI in 3D world
The UI systems and 3D world systems are now combined and function together.

Objects in the 3D world can be interacted with via clicking to open the storage UI and the associated pen's UI.

Player and Camera controls are disabled when inside of a UI, and re-enabled when exiting.

Objects that need to be limited by the presence of a UI can subscribe to the events OnUIOpened and OnUIClosed.